backend/requirements.txt
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
openai==1.3.6
deepseek-ai==0.1.2
coqui-tts==0.13.0
websockets==11.0.3
